ref: unify slurm and TE under backendPlugin 5/n" (#4582)
* ref: unify slurm and TE under backendPlugin 4/n

* ref: unify slurm and TE under backendPlugin 5/n
